tg_groups_api_old/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
telethon==1.42.0
python-dotenv==1.2.1
pydantic==2.12.5
prometheus-client==0.21.0
redis==5.2.0
